backend/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-community
langchain-google-genai
langchain-text-splitters

google-generativeai
chromadb
youtube-transcript-api
pydantic